Refactor File structure
Move indicators to folder
