+++ lib/ruby_technical_analysis/indicators/bollinger_bands.rb
