+++ lib/ruby_technical_analysis/indicators/chaikin_money_flow.rb
