+++ lib/ruby_technical_analysis/indicators/chande_momentum_oscillator.rb
