+++ lib/ruby_technical_analysis/indicators/commodity_channel_index.rb
