+++ lib/ruby_technical_analysis/indicators/envelopes_ema.rb
