+++ lib/ruby_technical_analysis/indicators/intraday_momentum_index.rb
