+++ lib/ruby_technical_analysis/indicators/macd.rb
