+++ lib/ruby_technical_analysis/indicators/mass_index.rb
