+++ lib/ruby_technical_analysis/indicators/moving_averages.rb
