+++ lib/ruby_technical_analysis/indicators/pivot_points.rb
