+++ lib/ruby_technical_analysis/indicators/price_channel.rb
