+++ lib/ruby_technical_analysis/indicators/qstick.rb
