+++ lib/ruby_technical_analysis/indicators/rate_of_change.rb
