+++ lib/ruby_technical_analysis/indicators/relative_momentum_index.rb
