+++ lib/ruby_technical_analysis/indicators/relative_strength_index.rb
