+++ lib/ruby_technical_analysis/indicators/statistical_methods.rb
