+++ lib/ruby_technical_analysis/indicators/stochastic_oscillator.rb
