+++ lib/ruby_technical_analysis/indicators/volume_oscillator.rb
