+++ lib/ruby_technical_analysis/indicators/volume_rate_of_change.rb
