+++ lib/ruby_technical_analysis/indicators/wilders_smoothing.rb
