+++ lib/ruby_technical_analysis/indicators/williams_percent_r.rb
